+++ src/main/java/de/srendi/advancedperipherals/common/container/InventoryManagerContainer.java
@@ -23,7 +23,7 @@ public class InventoryManagerContainer extends BaseContainer {
         layoutPlayerInventorySlots(7, 84);
         if (tileEntity != null) {
             tileEntity.getCapability(ForgeCapabilities.ITEM_HANDLER).ifPresent(handler -> {
-                addSlot(new SlotInputHandler(handler, 0, 79, 29, new SlotCondition().setNeededItem(Items.MEMORY_CARD.get()))); //Input
+                addSlot(new SlotInputHandler(handler, 0, SLOT_X, SLOT_Y, new SlotCondition().setNeededItem(Items.MEMORY_CARD.get()))); //Input
             });
         }
     }